[Feature/home_network] home/store, home/mypage 네트워크 작업 (#79)
* [feat/mypage]: feat Network MyPage

* [feat/home/mypage]: UI edit

* [feat/home]: UI bottomNavigation height and padding edit

* [feat/mypage]: chore string edit

* [feat/home/store]: UI albumCover detail

* [feat/store]: feat store network

* [feat/store]: feat 스와이프시 앨범 갯수 반영

* [feat/mypage]: chore remove blank

* [feat/store]: fix 서버에서 내려주는 변수 추가로 인한 DTO 수정

* [feat/mypage]: refactor store code

+++ app/src/main/java/com/teampophory/pophory/feature/home/mypage/MyPageDisplayItem.kt
@@ -17,7 +17,7 @@ data class MyPageInfo(
     val photos: List<Photo>
 ) {
     data class Photo(
-        val photoId: Long,
+        val photoId: Int,
         val photoUrl: String
     )
 }
